MB-16181: Change VB::Manifest stats to use fmt::memory_buffer
Switch from snprintf to fmtlib API

Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137969
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/collections/manifest.cc
@@ -295,7 +295,7 @@ nlohmann::json Manifest::toJson(
         }
         if (!scope["collections"].empty() || visible) {
             scope["name"] = s.second.name;
-            scope["uid"] = fmt::format("{0:x}", s.first);
+            scope["uid"] = fmt::format("{0:x}", uint32_t(s.first));
             manifest["scopes"].push_back(scope);
         }
     }
